gbm

Struct BufferObject

pub struct BufferObject<T: 'static> { /* private fields */ }

A GBM buffer object

Implementations

impl<T: 'static> BufferObject<T>

pub fn width(&self) -> u32

Get the width of the buffer object

pub fn height(&self) -> u32

Get the height of the buffer object

pub fn stride(&self) -> u32

Get the stride of the buffer object

pub fn stride_for_plane(&self, plane: i32) -> u32

Get the stride of the buffer object

pub fn format(&self) -> Format

Get the format of the buffer object

pub fn bpp(&self) -> u32

Get the bits per pixel of the buffer object

pub fn offset(&self, plane: i32) -> u32

Get the offset for a plane of the buffer object

pub fn plane_count(&self) -> u32

Get the plane count of the buffer object

pub fn modifier(&self) -> Modifier

Get the modifier of the buffer object

pub fn fd(&self) -> Result<OwnedFd, InvalidFdError>

Get a DMA-BUF file descriptor for the buffer object

This function creates a DMA-BUF (also known as PRIME) file descriptor handle for the buffer object. Each call to Self::fd() returns a new file descriptor and the caller is responsible for closing the file descriptor.

pub fn device_fd(&self) -> BorrowedFd<'_>

Get the file descriptor of the gbm device of this buffer object

pub fn handle(&self) -> BufferObjectHandle

Get the handle of the buffer object

This is stored in the platform generic union BufferObjectHandle type. However the format of this handle is platform specific.

pub fn fd_for_plane(&self, plane: i32) -> Result<OwnedFd, InvalidFdError>

Get a DMA-BUF file descriptor for a plane of the buffer object

This function creates a DMA-BUF (also known as PRIME) file descriptor handle for a plane of the buffer object. Each call to Self::fd_for_plane() returns a new file descriptor and the caller is responsible for closing the file descriptor.

pub fn handle_for_plane(&self, plane: i32) -> BufferObjectHandle

Get the handle of a plane of the buffer object

This is stored in the platform generic union BufferObjectHandle type. However the format of this handle is platform specific.

pub fn map<'a, F, S>( &'a self, x: u32, y: u32, width: u32, height: u32, f: F, ) -> IoResult<S>
where F: FnOnce(&MappedBufferObject<'a, T>) -> S,

Map a region of a GBM buffer object for cpu access

This function maps a region of a GBM bo for cpu read access.

pub fn map_mut<'a, F, S>( &'a mut self, x: u32, y: u32, width: u32, height: u32, f: F, ) -> IoResult<S>
where F: FnOnce(&mut MappedBufferObject<'a, T>) -> S,

Map a region of a GBM buffer object for cpu access

This function maps a region of a GBM bo for cpu read/write access.

pub fn write(&mut self, buffer: &[u8]) -> IoResult<()>

Write data into the buffer object

If the buffer object was created with the BufferObjectFlags::WRITE flag, this function can be used to write data into the buffer object. The data is copied directly into the object and it’s the responsibility of the caller to make sure the data represents valid pixel data, according to the width, height, stride and format of the buffer object.

pub fn set_userdata(&mut self, userdata: T) -> Option<T>

Sets the userdata of the buffer object.

If previously userdata was set, it is returned.

pub fn clear_userdata(&mut self)

Clears the set userdata of the buffer object.

pub fn userdata(&self) -> Option<&T>

Returns a reference to set userdata, if any.

pub fn userdata_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to set userdata, if any.

pub fn take_userdata(&mut self) -> Option<T>

Takes ownership of previously set userdata, if any.

This removes the userdata from the buffer object.