Struct gbm::MappedBufferObject

pub struct MappedBufferObject<'a, T: 'static> { /* private fields */ }

A mapped buffer object

Implementations

impl<'a, T: 'static> MappedBufferObject<'a, T>

pub fn stride(&self) -> u32

Get the stride of the buffer object

This is calculated by the backend when it does the allocation of the buffer.

pub fn height(&self) -> u32

The height of the mapped region for the buffer

pub fn width(&self) -> u32

The width of the mapped region for the buffer

pub fn x(&self) -> u32

The X (top left origin) starting position of the mapped region for the buffer

pub fn y(&self) -> u32

The Y (top left origin) starting position of the mapped region for the buffer

pub fn buffer(&self) -> &[u8]

Access to the underlying image buffer

pub fn buffer_mut(&mut self) -> &mut [u8]

Mutable access to the underlying image buffer