gbm

Struct Surface

pub struct Surface<T: 'static> { /* private fields */ }

A GBM rendering surface

Implementations

impl<T: 'static> Surface<T>

pub fn has_free_buffers(&self) -> bool

Return whether or not a surface has free (non-locked) buffers

Before starting a new frame, the surface must have a buffer available for rendering. Initially, a GBM surface will have a free buffer, but after one or more buffers have been locked, the application must check for a free buffer before rendering.

pub unsafe fn lock_front_buffer( &self, ) -> Result<BufferObject<T>, FrontBufferError>

Lock the surface’s current front buffer

Locks rendering to the surface’s current front buffer and returns a handle to the underlying BufferObject.

If an error occurs a FrontBufferError is returned.

Safety

This function must be called exactly once after calling eglSwapBuffers. Calling it before any eglSwapBuffers has happened on the surface or two or more times after eglSwapBuffers is an error and may cause undefined behavior.