Crate gbm

Safe libgbm bindings for rust

The Generic Buffer Manager

This module provides an abstraction that the caller can use to request a buffer from the underlying memory management system for the platform.

This allows the creation of portable code whilst still allowing access to the underlying memory manager.

This library is best used in combination with drm-rs, provided through the drm-support feature.

Example

extern crate drm;
extern crate gbm;

use drm::control::{self, crtc, framebuffer};
use gbm::{Device, Format, BufferObjectFlags};

// ... init your drm device ...
let drm = init_drm_device();

// init a gbm device
let gbm = Device::new(drm).unwrap();

// create a 4x4 buffer
let mut bo = gbm.create_buffer_object::<()>(
            1280, 720,
            Format::ARGB8888,
            BufferObjectFlags::SCANOUT | BufferObjectFlags::WRITE,
            ).unwrap();

// write something to it (usually use import or egl rendering instead)
let buffer = {
    let mut buffer = Vec::new();
    for i in 0..1280 {
        for _ in 0..720 {
            buffer.push(if i % 2 == 0 { 0 } else { 255 });
        }
    }
    buffer
};
bo.write(&buffer).unwrap();

// create a framebuffer from our buffer
let fb = gbm.add_framebuffer(&bo).unwrap();

// display it (and get a crtc, mode and connector before)
gbm.set_crtc(crtc_handle, Some(fb), (0, 0), &[con], Some(mode)).unwrap();

Structs

BufferObject
A gbm buffer object

BufferObjectFlags
Flags to indicate the intended use for the buffer - these are passed into Device::create_buffer_object.

Device
An open GBM device

DeviceDestroyedError
Thrown when the underlying gbm device was already destroyed

FdWrapper
Type wrapping a foreign file destructor

MappedBufferObject
A mapped buffer object

Surface
A gbm rendering surface

WrongDeviceError
Thrown when the gbm device does not belong to the buffer object

Enums

Format
Possible pixel formats used

FrontBufferError
Errors that may happen when locking the front buffer

Traits

AsRaw
Trait for types that allow to optain the underlying raw libinput pointer.

Type Definitions

BufferObjectHandle
Abstraction representing the handle to a buffer allocated by the manager