Struct gbm::Device

pub struct Device<T: AsRawFd + 'static> { /* fields omitted */ }

An open GBM device

Implementations

impl Device<FdWrapper>

pub unsafe fn new_from_fd(fd: RawFd) -> IoResult<Device<FdWrapper>>

Open a GBM device from a given unix file descriptor.

The file descriptor passed in is used by the backend to communicate with platform for allocating the memory. For allocations using DRI this would be the file descriptor returned when opening a device such as /dev/dri/card0.

Unsafety

The lifetime of the resulting device depends on the ownership of the file descriptor. Closing the file descriptor before dropping the Device will lead to undefined behavior.

impl<T: AsRawFd + 'static> Device<T>

pub fn new(fd: T) -> IoResult<Device<T>>

Open a GBM device from a given open DRM device.

The underlying file descriptor passed in is used by the backend to communicate with platform for allocating the memory. For allocations using DRI this would be the file descriptor returned when opening a device such as /dev/dri/card0.

pub fn backend_name(&self) -> &str

Get the backend name

pub fn is_format_supported(
&self,
format: Format,
usage: BufferObjectFlags
) -> bool

Test if a format is supported for a given set of usage flags

pub fn create_surface<U: 'static>(
&self,
width: u32,
height: u32,
format: Format,
usage: BufferObjectFlags
) -> IoResult<Surface<U>>

Allocate a new surface object

pub fn create_buffer_object<U: 'static>(
&self,
width: u32,
height: u32,
format: Format,
usage: BufferObjectFlags
) -> IoResult<BufferObject<U>>

Allocate a buffer object for the given dimensions

pub fn import_buffer_object_from_dma_buf<U: 'static>(
&self,
buffer: RawFd,
width: u32,
height: u32,
stride: u32,
format: Format,
usage: BufferObjectFlags
) -> IoResult<BufferObject<U>>

Create a gbm buffer object from an dma buffer

This function imports a foreign dma buffer from an open file descriptor and creates a new gbm buffer object for it. This enabled using the foreign object with a display API such as KMS.

The gbm bo shares the underlying pixels but its life-time is independent of the foreign object.